Split separate MoE compressed tensor schemes into their own files as is done for linear schemes and in VLLM.
